daemons: Add `dep` metatdata for dependency management.
Dep (https://github.com/golang/dep) is the most official looking
dependency management tool that I could find, and appears to do the
right thing most of the time.

Note: Since the status of the `client` daemon is in limbo right now,
it is explicitly excluded from dependency management.

Part of #1.